backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.9.1
torchvision==0.24.1
transformers==4.35.0
Pillow>=10.2.0
numpy>=1.26.0

